GraphQLAPI_SMS/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
strawberry-graphql
Flask-graphql
